requirements.txt
azure-functions
joblib
pandas
scikit-learn
